fix circular imports/bad path refactors
