+++ great_expectations/datasource/fluent/data_asset/path/directory_asset.py
@@ -9,10 +9,10 @@ from great_expectations.core import IDDict
 from great_expectations.core.batch import LegacyBatchDefinition
 from great_expectations.core.partitioners import Partitioner
 from great_expectations.datasource.fluent.constants import _DATA_CONNECTOR_NAME
-from great_expectations.datasource.fluent.data_asset.data_connector import FILE_PATH_BATCH_SPEC_KEY
 from great_expectations.datasource.fluent.data_asset.path.file_path_data_asset import (
     _FilePathDataAsset,
 )
+from great_expectations.datasource.fluent.data_connector import FILE_PATH_BATCH_SPEC_KEY
 from great_expectations.datasource.fluent.interfaces import DatasourceT
 
 if TYPE_CHECKING:
