+++ great_expectations/datasource/fluent/data_asset/path/regex_asset.py
@@ -12,11 +12,11 @@ from great_expectations.core.partitioners import (
     PartitionerYearly,
     RegexPartitioner,
 )
-from great_expectations.datasource.fluent.data_asset.data_connector import FILE_PATH_BATCH_SPEC_KEY
-from great_expectations.datasource.fluent.data_asset.data_connector.regex_parser import RegExParser
 from great_expectations.datasource.fluent.data_asset.path.file_path_data_asset import (
     _FilePathDataAsset,
 )
+from great_expectations.datasource.fluent.data_connector import FILE_PATH_BATCH_SPEC_KEY
+from great_expectations.datasource.fluent.data_connector.regex_parser import RegExParser
 from great_expectations.datasource.fluent.interfaces import DatasourceT
 
 if TYPE_CHECKING:
