+++ great_expectations/datasource/fluent/data_asset/path/spark/orc_asset.py
@@ -4,13 +4,12 @@ from typing import Literal, Optional, Union
 
 from pydantic import Field
 
-from great_expectations.compatibility.pydantic import pydantic
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.datasource.fluent.data_asset.path.directory_asset import (
     DirectoryDataAsset,
 )
 from great_expectations.datasource.fluent.data_asset.path.regex_asset import RegexDataAsset
-from great_expectations.datasource.fluent.data_asset.path.spark.spark_asset import (
+from great_expectations.datasource.fluent.data_asset.path.spark.spark_generic import (
     _SparkGenericFilePathAssetMixin,
 )
 
@@ -21,7 +20,8 @@ class ORCAssetBase(_SparkGenericFilePathAssetMixin):
     merge_schema: Optional[Union[bool, str]] = Field(False, alias="mergeSchema")
 
     class Config:
-        extra = pydantic.Extra.forbid
+        extra = "forbid"
+
         allow_population_by_field_name = True
 
     @classmethod
