+++ great_expectations/datasource/fluent/data_asset/path/spark/text_asset.py
@@ -4,13 +4,12 @@ from typing import Literal, Optional
 
 from pydantic import Field
 
-from great_expectations.compatibility.pydantic import pydantic
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.datasource.fluent.data_asset.path.directory_asset import (
     DirectoryDataAsset,
 )
 from great_expectations.datasource.fluent.data_asset.path.regex_asset import RegexDataAsset
-from great_expectations.datasource.fluent.data_asset.path.spark.spark_asset import (
+from great_expectations.datasource.fluent.data_asset.path.spark.spark_generic import (
     _SparkGenericFilePathAssetMixin,
 )
 
@@ -22,7 +21,8 @@ class TextAssetBase(_SparkGenericFilePathAssetMixin):
     line_sep: Optional[str] = Field(None, alias="lineSep")
 
     class Config:
-        extra = pydantic.Extra.forbid
+        extra = "forbid"
+
         allow_population_by_field_name = True
 
     @classmethod
