+++ great_expectations/datasource/fluent/data_connector/azure_blob_storage_data_connector.py
@@ -11,7 +11,7 @@ from great_expectations.datasource.data_connector.util import (
     list_azure_keys,
     sanitize_prefix,
 )
-from great_expectations.datasource.fluent.data_asset.data_connector import (
+from great_expectations.datasource.fluent.data_connector import (
     FilePathDataConnector,
 )
