+++ great_expectations/datasource/fluent/data_connector/dbfs_data_connector.py
@@ -6,7 +6,7 @@ import re
 from typing import TYPE_CHECKING, Callable, Optional
 
 from great_expectations.compatibility.typing_extensions import override
-from great_expectations.datasource.fluent.data_asset.data_connector import (
+from great_expectations.datasource.fluent.data_connector import (
     FilesystemDataConnector,
 )
