+++ great_expectations/datasource/fluent/data_connector/filesystem_data_connector.py
@@ -11,7 +11,7 @@ from great_expectations.datasource.data_connector.util import (
     get_filesystem_one_level_directory_glob_path_list,
     normalize_directory_path,
 )
-from great_expectations.datasource.fluent.data_asset.data_connector import (
+from great_expectations.datasource.fluent.data_connector import (
     FilePathDataConnector,
 )
