+++ great_expectations/datasource/fluent/spark_dbfs_datasource.py
@@ -7,7 +7,7 @@ from great_expectations._docs_decorators import public_api
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.core.util import DBFSPath
 from great_expectations.datasource.fluent import SparkFilesystemDatasource
-from great_expectations.datasource.fluent.data_asset.data_connector import (
+from great_expectations.datasource.fluent.data_connector import (
     DBFSDataConnector,
 )
