+++ tests/datasource/fluent/data_asset/data_connector/test_s3_data_connector.py
@@ -16,14 +16,14 @@ from great_expectations.datasource.data_connector.util import (
     sanitize_prefix_for_gcs_and_s3,
 )
 from great_expectations.datasource.fluent import BatchRequest
-from great_expectations.datasource.fluent.data_asset.data_connector import (
+from great_expectations.datasource.fluent.data_connector import (
     S3DataConnector,
 )
 
 if TYPE_CHECKING:
     from botocore.client import BaseClient
 
-    from great_expectations.datasource.fluent.data_asset.data_connector import (
+    from great_expectations.datasource.fluent.data_connector import (
         DataConnector,
     )
