+++ tests/datasource/fluent/test_spark_s3_datasource.py
@@ -10,10 +10,10 @@ import pytest
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core.util import S3Url
 from great_expectations.datasource.fluent import SparkS3Datasource
-from great_expectations.datasource.fluent.data_asset.data_connector import (
+from great_expectations.datasource.fluent.data_asset.path.spark.csv_asset import CSVAsset
+from great_expectations.datasource.fluent.data_connector import (
     S3DataConnector,
 )
-from great_expectations.datasource.fluent.data_asset.path.spark.csv_asset import CSVAsset
 from great_expectations.datasource.fluent.file_path_data_asset import (
     _FilePathDataAsset,
 )
